Improve the speed of getting file sizes

+++ sdks/python/apache_beam/io/filebasedsource.py
@@ -107,7 +107,8 @@ class FileBasedSource(iobase.BoundedSource):
     if self._concat_source is None:
       single_file_sources = []
       file_names = [f for f in fileio.ChannelFactory.glob(self._pattern)]
-      sizes = FileBasedSource._estimate_sizes_in_parallel(file_names)
+      sizes = FileBasedSource._estimate_sizes_of_files(file_names,
+                                                       self._pattern)
 
       # We create a reference for FileBasedSource that will be serialized along
       # with each _SingleFileSource. To prevent this FileBasedSource from having
@@ -144,22 +145,32 @@ class FileBasedSource(iobase.BoundedSource):
         compression_type=self._compression_type)
 
   @staticmethod
-  def _estimate_sizes_in_parallel(file_names):
+  def _estimate_sizes_of_files(file_names, pattern=None):
+    """Returns the size of all the files as an ordered list based on the file
+    names that are provided here. If the pattern is specified here then we use
+    the size_of_files_in_glob method to get the size of files matching the glob
+    for performance improvements instead of getting the size one by one.
+    """
     if not file_names:
       return []
     elif len(file_names) == 1:
       return [fileio.ChannelFactory.size_in_bytes(file_names[0])]
     else:
-      # ThreadPool crashes in old versions of Python (< 2.7.5) if created from a
-      # child thread. (http://bugs.python.org/issue10015)
-      if not hasattr(threading.current_thread(), '_children'):
-        threading.current_thread()._children = weakref.WeakKeyDictionary()
-      pool = ThreadPool(
-          min(MAX_NUM_THREADS_FOR_SIZE_ESTIMATION, len(file_names)))
-      try:
-        return pool.map(fileio.ChannelFactory.size_in_bytes, file_names)
-      finally:
-        pool.terminate()
+      if pattern is None:
+        # ThreadPool crashes in old versions of Python (< 2.7.5) if created
+        # from a child thread. (http://bugs.python.org/issue10015)
+        if not hasattr(threading.current_thread(), '_children'):
+          threading.current_thread()._children = weakref.WeakKeyDictionary()
+        pool = ThreadPool(
+            min(MAX_NUM_THREADS_FOR_SIZE_ESTIMATION, len(file_names)))
+        try:
+          return pool.map(fileio.ChannelFactory.size_in_bytes, file_names)
+        finally:
+          pool.terminate()
+      else:
+        file_sizes = fileio.ChannelFactory.size_of_files_in_glob(pattern,
+                                                                 file_names)
+        return [file_sizes[f] for f in file_names]
 
   def _validate(self):
     """Validate if there are actual files in the specified glob pattern
@@ -179,7 +190,10 @@ class FileBasedSource(iobase.BoundedSource):
     file_names = [f for f in fileio.ChannelFactory.glob(self._pattern)]
     if (len(file_names) <=
         FileBasedSource.MIN_NUMBER_OF_FILES_TO_STAT):
-      return sum(self._estimate_sizes_in_parallel(file_names))
+      # We're reading very few files so we can pass names without pattern
+      # as otherwise we'll try to do optimization based on the pattern and
+      # might end up reading much more data than needed for a few files.
+      return sum(self._estimate_sizes_of_files(file_names))
     else:
       # Estimating size of a random sample.
       # TODO: better support distributions where file sizes are not
@@ -188,7 +202,7 @@ class FileBasedSource(iobase.BoundedSource):
                         int(len(file_names) *
                             FileBasedSource.MIN_FRACTION_OF_FILES_TO_STAT))
       sample = random.sample(file_names, sample_size)
-      estimate = self._estimate_sizes_in_parallel(sample)
+      estimate = self._estimate_sizes_of_files(sample, self._pattern)
       return int(
           sum(estimate) *
           (float(len(file_names)) / len(sample)))
